fix octane global configuration
