+++ src/main/resources/com/microfocus/application/automation/tools/settings/OctaneServerSettingsGlobalConfiguration/config.properties
