+++ src/main/resources/com/microfocus/application/automation/tools/settings/OctaneServerSettingsGlobalConfiguration/help-sscBaseToken.html
